Add clear_inputs.js, get_scores.js, post_score.js and render.js files
