+++ src/index.html
@@ -13,14 +13,14 @@
   <main>
     <section>
       <h2>Recent scores <button class="border" id="refresh-button">Refresh</button></h2>
-      <ul id="scores-list" class="border"></ul>
+      <ul class="border" id="scores-list"></ul>
     </section>
     <section>
       <h2>Add your score</h2>
       <form>
-        <input id="name" class="border" type="text" placeholder="Your name" required>
-        <input id="number" class="border" type="number" placeholder="Your score" required>
-        <div class="submit-button-div">
+        <input class="border" id="name" type="text" placeholder="Your name" required>
+        <input class="border"id="number" type="number" placeholder="Your score" required>
+        <div class="submit-button-container">
           <button class="border">Submit</button>
         </div>
       </form>
