+++ src/style.css
@@ -34,6 +34,10 @@ input {
   margin-bottom: 10%;
 }
 
-.submit-button-div {
+.submit-button-container {
   align-self: end;
+}
+
+button {
+  cursor: pointer;
 }
